+++ WebContent/telefones.jsp
@@ -80,7 +80,7 @@
 						<td data-title="Tipo">
 							<c:out value="${fone.tipo}"></c:out>
 						</td>
-						<td data-title="Excluir"><a href="salvarTelefones?user=${fone.usuario}&acao=deletefone&foneId=${fone.id}"> <img
+						<td data-title="Excluir"><a href="salvarTelefones?user=${fone.usuario}&acao=deletefone&foneId=${fone.id}" onclick="return confirm('Confirmar a exclusão?')"> <img
 									src="resources/img/icon.png" width="20px" height="20px" title="Excluir"
 									alt="Excluir">
 							</a></td>
